refactor: curdImpl ,now code style more like OOP rather than FP
