+++ src/main/scala/com/parsley/orm/curd/update/UpdateOperation.scala
@@ -1,4 +1,4 @@
-package com.parsley.orm
+package com.parsley.orm.curd.update
 
 class UpdateOperation {
 
@@ -15,14 +15,6 @@ class UpdateOperation {
 }
 
 object UpdateOperation {
-
-    extension (self: String) {
-        def ==> (x: Int | Double | String | Long | Boolean | Float | Char): UpdateOperation = {
-            val res = new UpdateOperation
-            res.sqlString.append(s"`$self`='${x.toString}'")
-            res
-        }
-    }
 
 }
